requirements.txt
langgraph
groq
pdfminer
pinecone-client
sentence_transformers